+++ src/main/java/net/snakefangox/worldshell/storage/LocalSpace.java
@@ -1,10 +1,9 @@
 package net.snakefangox.worldshell.storage;
 
-import com.jme3.math.Quaternion;
-import com.jme3.math.Vector3f;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3d;
-import net.snakefangox.worldshell.collision.ShellCollisionHull;
+import net.snakefangox.worldshell.math.Quaternion;
+import net.snakefangox.worldshell.math.Vector3d;
 
 /**
  * Indicates that a given object contains a local coordinate space that differs from
@@ -89,145 +88,145 @@ public interface LocalSpace {
 	/** -- Below this line lies only madness and a lot of matrix transforms -- **/
 
 	default BlockPos toLocal(BlockPos pos) {
-		float newX = (float) (pos.getX() - getLocalX());
-		float newY = (float) (pos.getY() - getLocalY());
-		float newZ = (float) (pos.getZ() - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (pos.getX() - getLocalX());
+		double newY = (pos.getY() - getLocalY());
+		double newZ = (pos.getZ() - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		getInverseRotation().multLocal(vec);
 		return new BlockPos(vec.x, vec.y, vec.z);
 	}
 
 	default BlockPos toGlobal(BlockPos pos) {
-		Vector3f vec = new Vector3f(pos.getX(), pos.getY(), pos.getZ());
+		Vector3d vec = new Vector3d(pos.getX(), pos.getY(), pos.getZ());
 		getRotation().multLocal(vec);
 		return new BlockPos(vec.x + getLocalX(), vec.y + getLocalY(), vec.z + getLocalZ());
 	}
 
 	default Vec3d toLocal(Vec3d pos) {
-		float newX = (float) (pos.getX() - getLocalX());
-		float newY = (float) (pos.getY() - getLocalY());
-		float newZ = (float) (pos.getZ() - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (pos.getX() - getLocalX());
+		double newY = (pos.getY() - getLocalY());
+		double newZ = (pos.getZ() - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		getInverseRotation().multLocal(vec);
 		return new Vec3d(vec.x, vec.y, vec.z);
 	}
 
 	default Vec3d toGlobal(Vec3d pos) {
-		Vector3f vec = new Vector3f((float) pos.getX(), (float) pos.getY(), (float) pos.getZ());
+		Vector3d vec = new Vector3d(pos.getX(), pos.getY(), pos.getZ());
 		getRotation().multLocal(vec);
 		return new Vec3d(vec.x + getLocalX(), vec.y + getLocalY(), vec.z + getLocalZ());
 	}
 
 	default Vec3d toLocal(double x, double y, double z) {
-		float newX = (float) (x - getLocalX());
-		float newY = (float) (y - getLocalY());
-		float newZ = (float) (z - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (x - getLocalX());
+		double newY = (y - getLocalY());
+		double newZ = (z - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		getInverseRotation().multLocal(vec);
 		return new Vec3d(vec.x, vec.y, vec.z);
 	}
 
 	default Vec3d toGlobal(double x, double y, double z) {
-		Vector3f vec = new Vector3f((float) x, (float) y, (float) z);
+		Vector3d vec = new Vector3d(x, y, z);
 		getRotation().multLocal(vec);
 		return new Vec3d(vec.x + getLocalX(), vec.y + getLocalY(), vec.z + getLocalZ());
 	}
 
 	default BlockPos toLocal(int x, int y, int z) {
-		float newX = (float) (x - getLocalX());
-		float newY = (float) (y - getLocalY());
-		float newZ = (float) (z - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (x - getLocalX());
+		double newY = (y - getLocalY());
+		double newZ = (z - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		getInverseRotation().multLocal(vec);
 		return new BlockPos(vec.x, vec.y, vec.z);
 	}
 
 	default BlockPos toGlobal(int x, int y, int z) {
-		Vector3f vec = new Vector3f((float) x, (float) y, (float) z);
+		Vector3d vec = new Vector3d(x, y, z);
 		getRotation().multLocal(vec);
 		return new BlockPos(vec.x + getLocalX(), vec.y + getLocalY(), vec.z + getLocalZ());
 	}
 
 	default BlockPos.Mutable toLocal(BlockPos.Mutable pos) {
-		float newX = (float) (pos.getX() - getLocalX());
-		float newY = (float) (pos.getY() - getLocalY());
-		float newZ = (float) (pos.getZ() - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (pos.getX() - getLocalX());
+		double newY = (pos.getY() - getLocalY());
+		double newZ = (pos.getZ() - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		getInverseRotation().multLocal(vec);
 		return pos.set(vec.x, vec.y, vec.z);
 	}
 
 	default BlockPos.Mutable toGlobal(BlockPos.Mutable pos) {
-		Vector3f vec = new Vector3f(pos.getX(), pos.getY(), pos.getZ());
+		Vector3d vec = new Vector3d(pos.getX(), pos.getY(), pos.getZ());
 		getRotation().multLocal(vec);
 		return pos.set(vec.x + getLocalX(), vec.y + getLocalY(), vec.z + getLocalZ());
 	}
 
-	default Vector3f toLocal(Vector3f pos) {
-		float newX = (float) (pos.x - getLocalX());
-		float newY = (float) (pos.y - getLocalY());
-		float newZ = (float) (pos.z - getLocalZ());
+	default Vector3d toLocal(Vector3d pos) {
+		double newX = (pos.x - getLocalX());
+		double newY = (pos.y - getLocalY());
+		double newZ = (pos.z - getLocalZ());
 		return getInverseRotation().multLocal(pos.set(newX, newY, newZ));
 	}
 
-	default Vector3f toGlobal(Vector3f pos) {
+	default Vector3d toGlobal(Vector3d pos) {
 		getRotation().multLocal(pos);
-		return pos.addLocal((float) getLocalX(), (float) getLocalY(), (float) getLocalZ());
+		return pos.addLocal(getLocalX(), getLocalY(), getLocalZ());
 	}
 
 	default double toLocalX(double x, double y, double z) {
-		float newX = (float) (x - getLocalX());
-		float newY = (float) (y - getLocalY());
-		float newZ = (float) (z - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (x - getLocalX());
+		double newY = (y - getLocalY());
+		double newZ = (z - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		return getInverseRotation().multLocal(vec).x;
 	}
 
 	default double toLocalY(double x, double y, double z) {
-		float newX = (float) (x - getLocalX());
-		float newY = (float) (y - getLocalY());
-		float newZ = (float) (z - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (x - getLocalX());
+		double newY = (y - getLocalY());
+		double newZ = (z - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		return getInverseRotation().multLocal(vec).y;
 	}
 
 	default double toLocalZ(double x, double y, double z) {
-		float newX = (float) (x - getLocalX());
-		float newY = (float) (y - getLocalY());
-		float newZ = (float) (z - getLocalZ());
-		Vector3f vec = new Vector3f(newX, newY, newZ);
+		double newX = (x - getLocalX());
+		double newY = (y - getLocalY());
+		double newZ = (z - getLocalZ());
+		Vector3d vec = new Vector3d(newX, newY, newZ);
 		return getInverseRotation().multLocal(vec).z;
 	}
 
 	default double toGlobalX(double x, double y, double z) {
-		return getRotation().multLocal(new Vector3f((float) x, (float) y, (float) z)).x + getLocalX();
+		return getRotation().multLocal(new Vector3d(x, y, z)).x + getLocalX();
 	}
 
 	default double toGlobalY(double x, double y, double z) {
-		return getRotation().multLocal(new Vector3f((float) x, (float) y, (float) z)).y + getLocalY();
+		return getRotation().multLocal(new Vector3d(x, y, z)).y + getLocalY();
 	}
 
 	default double toGlobalZ(double x, double y, double z) {
-		return getRotation().multLocal(new Vector3f((float) x, (float) y, (float) z)).z + getLocalZ();
+		return getRotation().multLocal(new Vector3d(x, y, z)).z + getLocalZ();
 	}
 
 	default BlockPos globalToGlobal(LocalSpace target, BlockPos pos) {
-		Vector3f vec = toLocal(new Vector3f(pos.getX(), pos.getY(), pos.getZ()));
+		Vector3d vec = toLocal(new Vector3d(pos.getX(), pos.getY(), pos.getZ()));
 		return target.toGlobal((int) vec.x, (int) vec.y, (int) vec.z);
 	}
 
 	default Vec3d globalToGlobal(LocalSpace target, Vec3d pos) {
-		Vector3f vec = toLocal(new Vector3f((float) pos.x, (float) pos.y, (float) pos.z));
+		Vector3d vec = toLocal(new Vector3d(pos.x, pos.y, pos.z));
 		return target.toGlobal(vec.x, vec.y, vec.z);
 	}
 
 	default Vec3d globalToGlobal(LocalSpace target, double x, double y, double z) {
-		Vector3f vec = toLocal(new Vector3f((float) x, (float) y, (float) z));
+		Vector3d vec = toLocal(new Vector3d(x, y, z));
 		return target.toGlobal(vec.x, vec.y, vec.z);
 	}
 
 	default BlockPos globalToGlobal(LocalSpace target, int x, int y, int z) {
-		Vector3f vec = toLocal(new Vector3f(x, y, z));
+		Vector3d vec = toLocal(new Vector3d(x, y, z));
 		return target.toGlobal((int) vec.x, (int) vec.y, (int) vec.z);
 	}
 
@@ -236,17 +235,17 @@ public interface LocalSpace {
 	}
 
 	default double globalToGlobalX(LocalSpace target, double x, double y, double z) {
-		Vector3f vec = toLocal(new Vector3f((float) x, (float) y, (float) z));
+		Vector3d vec = toLocal(new Vector3d(x, y, z));
 		return target.toGlobalX(vec.x, vec.y, vec.z);
 	}
 
 	default double globalToGlobalY(LocalSpace target, double x, double y, double z) {
-		Vector3f vec = toLocal(new Vector3f((float) x, (float) y, (float) z));
+		Vector3d vec = toLocal(new Vector3d(x, y, z));
 		return target.toGlobalY(vec.x, vec.y, vec.z);
 	}
 
 	default double globalToGlobalZ(LocalSpace target, double x, double y, double z) {
-		Vector3f vec = toLocal(new Vector3f((float) x, (float) y, (float) z));
+		Vector3d vec = toLocal(new Vector3d(x, y, z));
 		return target.toGlobalZ(vec.x, vec.y, vec.z);
 	}
 }
